formatTime.jsx 생성
