+++ src/admin/ReportManagement.jsx
@@ -1,6 +1,7 @@
 import React, { useState, useEffect, useCallback } from 'react';
 import Swal from 'sweetalert2';
 import axios from 'axios';
+import formatTime from '../util/formatTime';
 
 const BoardReportCategory = {
   SPAM: '스팸홍보/도배글',
@@ -126,17 +127,6 @@ const ReportManagement = () => {
       icon: 'info',
       confirmButtonText: '확인'
     });
-  };
-
-  // 날짜 설정
-  const formatTime = (dateString) => {
-    const date = new Date(dateString);
-    const year = date.getFullYear();
-    const month = (date.getMonth() + 1).toString().padStart(2, '0');
-    const day = date.getDate().toString().padStart(2, '0');
-    const hours = date.getHours().toString().padStart(2, '0');
-    const minutes = date.getMinutes().toString().padStart(2, '0');
-    return `${year}-${month}-${day} / ${hours}:${minutes}`;
   };
 
   if (error) return <div>{error}</div>;
