+++ src/util/formatTime.jsx
@@ -1,0 +1,13 @@
+  // 날짜 설정
+
+const formatTime = (dateString) => {
+  const date = new Date(dateString);
+  const year = date.getFullYear();
+  const month = (date.getMonth() + 1).toString().padStart(2, "0");
+  const day = date.getDate().toString().padStart(2, "0");
+  const hours = date.getHours().toString().padStart(2, "0");
+  const minutes = date.getMinutes().toString().padStart(2, "0");
+  return `${year}-${month}-${day} / ${hours}:${minutes}`;
+};
+
+export default formatTime;
